Renaming motion metrics
