+++ svelte-web-frontend/src/lib/ai/motionmetrics/Julien2DSkeletonSimilarityMetric.ts
