+++ svelte-web-frontend/src/lib/ai/motionmetrics/MotionMetric.ts
