+++ svelte-web-frontend/src/lib/ai/motionmetrics/Skeleton3dVectorAngleSimilarityMetric.ts
